refactor: Setting 클래스를 도메인 모듈로 이동
